Pages added
GooPage added

BingPage added

rebase commits
