+++ src/test/java/com/berezkindv/tests/ParametrizedYandexTestWithValueSource.java
@@ -1,6 +1,6 @@
 package com.berezkindv.tests;
 
-import com.berezkindv.pages.TestPages;
+import com.berezkindv.pages.YaPage;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -9,12 +9,12 @@ import static com.berezkindv.tests.TestData.yaSearch;
 
 public class ParametrizedYandexTestWithValueSource {
 
-    TestPages testPages = new TestPages();
+    YaPage yaPage = new YaPage();
 
     @ValueSource(strings = {"Selenide", "Allure", "Junit"})
     @ParameterizedTest(name = "Тестирование общего алгоритма поиска с тестовыми данными: {0}")
     void commonSearchTest(String testData) {
-        testPages.openPage(yaSearch)
+        yaPage.openPage(yaSearch)
                 .typeYaSearchData(testData)
                 .pushYaSubmitButton()
                 .checkYaResult(testData);
